backend/requirements.txt
fastapi==0.110.0
uvicorn>=0.30.0
google-genai==1.3.0
python-dotenv==1.0.1
python-multipart==0.0.9
pillow==10.2.0
